+++ src/drawio_exporter/core/drawio/drawio_desktop.rs
@@ -87,6 +87,7 @@ pub struct ExportArguments<'a> {
     pub all_pages: bool,
     pub page_index: Option<&'a str>,
     pub page_range: Option<&'a str>,
+    pub embed_svg_images: bool,
 }
 
 impl<'a> ExportArguments<'a> {
@@ -155,6 +156,10 @@ impl<'a> ExportArguments<'a> {
 
         if self.uncompressed {
             arguments.push("--uncompressed");
+        }
+
+        if self.embed_svg_images {
+            arguments.push("--embed-svg-images");
         }
 
         // Input is always the last argument